+++ R/tt_dotabulation.R
@@ -892,6 +892,7 @@ build_table = function(lyt, df,
                        col_counts = NULL,
                        col_total = if(is.null(alt_counts_df)) nrow(df) else nrow(alt_counts_df),
                        topleft = NULL,
+                       hdr_sep = .default_hsep(),
                        ...) {
     if(!is(lyt, "PreDataTableLayouts")) {
         stop("lyt must be a PreDataTableLayouts object. Got object of class ", class(lyt))
@@ -985,6 +986,7 @@ build_table = function(lyt, df,
     ## but now just call it so the error gets thrown when I want it to
     unused <- matrix_form(tab)
     tab <- update_ref_indexing(tab)
+    header_sep(tab) <- hdr_sep
     tab
 }